LayoutTests/imported/w3c/web-platform-tests/svg/types/scripted/SVGAnimatedInteger-expected.txt


FAIL SVGAnimatedInteger interface - utilizing the targetX property of SVGFEConvolveMatrix assert_equals: expected 3 but got -1
